Only refresh feeds when showing their articles.
Previously, we considered refreshing stale feeds every time the
FeedInfo was read from the datastore. This was unnecessary.

Also, when adding a feed, it was refreshed. But, if the refresh failed
then the feed was never subscribed, and it's reference count was never
incremented even though the FeedInfo was put in the datastore. Now, we
don't refresh the feed until we actually want it's articles, so this
issues goes away.

Fixes #63