Save article's publish dates because these are used in their keys.
We use the date in the key in case there are multiple articles with
the same title. But, we can't use the When field any more, because
it's not unique to the article, it is now the time when the article
was fetched. So, instead, we need to save the PubDate too.